Move pre/post-deployment script to Impl layer
